bezier pow3 method of interpolation

+++ barray/straight.hpp
@@ -8,10 +8,10 @@ data_T barray <data_T> :: operator [] (double index)
 {
     if (index >= 0 && index < _capacity)
     {
-        int     P0 = (int)index;
+        int     P0 = index;
         int     P1 = P0 + 1;
         double  t = index - P0;
-        return  _data[P0] * (1 - t)
+        return    _data[P0] * (1 - t)
                 + _data[P1] * t;
     }
 }
